+++ RELEASE-NOTES.md
@@ -82,6 +82,7 @@
 1. SQL Parser: Support Hive CREATE MATERIALIZED VIEW statement parse - [#36144](https://github.com/apache/shardingsphere/pull/36144)
 1. SQL Parser: Support Hive DROP MATERIALIZED VIEW statement parse - [#36145](https://github.com/apache/shardingsphere/pull/36145)
 1. SQL Parser: Support Hive ALTER MATERIALIZED VIEW statement parse - [#36146](https://github.com/apache/shardingsphere/pull/36146)
+1. SQL Parser: Support Hive CREATE INDEX statement parse - [#36157](https://github.com/apache/shardingsphere/pull/36157)
 1. SQL Parser: Support SQL Server xml methods parse - [#35911](https://github.com/apache/shardingsphere/pull/35911)
 1. SQL Parser: Support SQL Server CHANGETABLE function parse - [#35920](https://github.com/apache/shardingsphere/pull/35920)
 1. SQL Parser: Support SQL Server AI_GENERATE_EMBEDDINGS function parse - [#35922](https://github.com/apache/shardingsphere/pull/35922)